ml-service/requirements.txt
flask==2.3.3
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1
nltk==3.8.1
spacy==3.6.0
gensim==4.3.1
pymongo==4.4.1
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
joblib==1.3.1

